// src/data/types.ts
export type DayKey = "D" | "L" | "M" | "X" | "J" | "V" | "S";
// ...
export interface ClassBlock {
  id: string;
  courseName: string;
  professor: string;
  room: string;
  startTime: string; // HH:mm
  endTime: string; // HH:mm
}

export type WeeklySchedule = Record<DayKey, ClassBlock[]>;